style   : :high_heel: label-editor css 수정
