+++ web/client/src/components/label-editor/index.tsx
@@ -11,16 +11,20 @@ export default function LabelEditor() {
         <S.colorTitle>Color</S.colorTitle>
       </S.LabelTitleRow>
       <S.LableInputRow>
-        <S.nameInput />
-        <S.descriptionInput />
-        <S.colorInputContainer>
-          <S.IconContainer>
-            <S.refreshIcon />
-          </S.IconContainer>
-          <S.colorInput />
-        </S.colorInputContainer>
-        <Button value="cancel" color="white" />
-        <Button value="Create Name" color="green" />
+        <S.inputContainer>
+          <S.nameInput />
+          <S.descriptionInput />
+          <S.colorInputContainer>
+            <S.IconContainer>
+              <S.refreshIcon />
+            </S.IconContainer>
+            <S.colorInput />
+          </S.colorInputContainer>
+        </S.inputContainer>
+        <S.ButtonContainer>
+          <Button value="cancel" color="white" />
+          <Button value="Create Label" color="green" />
+        </S.ButtonContainer>
       </S.LableInputRow>
     </S.LabelEditor>
   );
