--- web/client/src/components/label-row-container/style.tsx
@@ -1,21 +1,0 @@
-import styled from "styled-components";
-
-export const LabelRowContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const LabelContainerHeader = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  background-color: #f5f8fa;
-  width: 65%;
-  padding: 0.7rem 1rem 0.7rem 1.5rem;
-  border-left: 1px solid #e8e8e8;
-  border-right: 1px solid #e8e8e8;
-  border-bottom: 1px solid #e8e8e8;
-`;
