--- web/client/src/views/label.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import LabelRow from "../components/label-row";
-import LabelRowContainer from "../components/label-row-container";
-
-const LabelPage = () => {
-  return (
-    <>
-      <LabelRowContainer />
-    </>
-  );
-};
-
-export default LabelPage;
