+++ web/client/src/views/label/index.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import LabelRowContainer from "../../containers/label-row-container";
+import LabelListHeader from "../../containers/label-list-header";
+
+const LabelPage = () => {
+  return (
+    <>
+      <LabelListHeader />
+      <LabelRowContainer />
+    </>
+  );
+};
+
+export default LabelPage;
